navigation-api/navigate-event/navigate-svg-anchor-download.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<svg><a id="a" href="foo.html"></a></svg>
<script>
for (const download_attr of ["", "filename"]) {
  async_test(t => {
    a.download = download_attr;
    navigation.onnavigate = t.step_func_done(e => {
      assert_equals(e.navigationType, "push");
      assert_true(e.cancelable);
      assert_true(e.canIntercept);
      assert_false(e.userInitiated);
      assert_false(e.hashChange);
      assert_equals(e.downloadRequest, download_attr);
      assert_equals(e.formData, null);
      assert_equals(new URL(e.destination.url).pathname,
                    "/navigation-api/navigate-event/foo.html");
      assert_false(e.destination.sameDocument);
      assert_equals(e.destination.key, "");
      assert_equals(e.destination.id, "");
      assert_equals(e.destination.index, -1);
      assert_equals(e.sourceElement, a);
      e.preventDefault();
    });
    a.dispatchEvent(new MouseEvent('click')); // SVG <a> doesn't support .click()
  }, `SVG <a> fires navigate and populates downloadRequest with '${download_attr}'`);
}
</script>
</body>
